Move git and touch helper functions to separate file
