+++ test/utils.py
@@ -1,0 +1,9 @@
+from stograde.common import run
+
+
+def git(cmd, *args):
+    return run(['git', cmd, *args])
+
+
+def touch(file):
+    return run(['touch', file])
